WIP: Start redesigning reading from sources
