+++ tests/test_concrete_settings.py
@@ -312,51 +312,3 @@ def test_nested_triple_nested_validation_errors():
         "Expected value of type `<class 'str'>` got value of type `<class 'int'>`"
     ]}]}]}
     # fmt: on
-
-
-#
-# Settings.update()
-#
-
-
-def test_update_empty_dict():
-    class S(Settings):
-        pass
-
-    S().update({})
-
-
-def test_update_with_extra_data_has_no_effect_on_settings():
-    class S(Settings):
-        pass
-
-    s = S()
-    s.update({'x': 10})
-    assert not hasattr(s, 'x')
-
-
-def test_update_top_level_setting():
-    class S(Settings):
-        T: int = 10
-
-    s = S()
-    s.update({'T': 100})
-    assert s.T == 100
-
-
-def test_update_nested_setting():
-    class S(Settings):
-        T: int = 10
-
-    class S1(Settings):
-        NESTED_S = S()
-
-    s1 = S1()
-    # fmt: off
-    s1.update({
-        'NESTED_S': {
-            'T': 20
-        }
-    })
-    # fmt: on
-    assert s1.NESTED_S.T == 20
